@@ -145,19 +147,19 @@
     "jasmine-spec-reporter": "5.0.2",
     "karma": "6.3.17",
     "karma-chrome-launcher": "3.1.0",
-    "karma-coverage-istanbul-reporter": "3.0.3",
+    "karma-coverage": "~2.2.1",
     "karma-jasmine": "4.0.1",
     "karma-jasmine-html-reporter": "1.7.0",
     "ngx-pipes": "3.0.0",
     "nodecat": "2.0.0",
+    "nx": "14.4.6",
     "nz-tslint-rules": "0.901.2",
-    "prettier": "2.8.0",
+    "prettier": "2.8.8",
     "prettier-eslint-cli": "7.1.0",
-    "protractor": "7.0.0",
     "rxjs-marbles": "7.0.1",
     "sass": "1.49.9",
     "ts-node": "4.1.0",
-    "typescript": "4.5.5",
+    "typescript": "4.6.4",
     "webpack-bundle-analyzer": "4.5.0"
   },
   "browserslist": [
