@@ -15,20 +15,20 @@ class TestProxyClient:
 
     @pytest.fixture
     def server(self):
-        server = ProxyServer()
+        server = ProxyServer(port=5005)
         yield server
-        server.shutdown()
+        server.graceful_shutdown()
 
     @pytest.fixture
     def server_with_dp(self, data_queue):
-        server = ProxyServer()
+        server = ProxyServer(port=5005)
         server.register_data_handler(
             lambda _, table: list(
                 map(data_queue.put, map(lambda t: t[1], table.to_pandas().iterrows()))
             )
         )
         yield server
-        server.shutdown()
+        server.graceful_shutdown()
 
     @pytest.fixture
     def client(self):
