@@ -2,6 +2,7 @@ from collections import deque
 
 import pytest
 
+from core.models.table import all_output_to_tuple
 from pytexera import Tuple
 from .echo_table_operator import EchoTableOperator
 
@@ -17,7 +18,7 @@ class TestEchoTableOperator:
         print(tuple_)
         deque(echo_table_operator.process_tuple(tuple_, 0))
         outputs = echo_table_operator.on_finish(0)
-        output_tuple = Tuple(next(outputs))
+        output_tuple = next(all_output_to_tuple(next(outputs)))
         assert output_tuple == tuple_
         with pytest.raises(StopIteration):
             next(outputs)
