@@ -22,9 +22,9 @@ def init_loguru_logger(stream_log_level) -> None:
 
 
 if __name__ == "__main__":
-
-    init_loguru_logger(sys.argv[3])
+    _, worker_id, output_port, logger_level = sys.argv
+    init_loguru_logger(logger_level)
 
     PythonWorker(
-        host="localhost", input_port=int(sys.argv[1]), output_port=int(sys.argv[2])
+        worker_id=worker_id, host="localhost", output_port=int(output_port)
     ).run()
