@@ -13,13 +13,22 @@ import org.apache.arrow.flight._
 import org.apache.arrow.memory.{BufferAllocator, RootAllocator}
 import org.apache.arrow.util.AutoCloseables
 
+import java.io.IOException
+import java.net.ServerSocket
+import java.util.concurrent.atomic.AtomicInteger
 import scala.collection.mutable
+import com.twitter.util.Promise
+
+import java.nio.charset.Charset
 
 private class AmberProducer(
     controlOutputPort: NetworkOutputPort[ControlPayload],
-    dataOutputPort: NetworkOutputPort[DataPayload]
+    dataOutputPort: NetworkOutputPort[DataPayload],
+    promise: Promise[Int]
 ) extends NoOpFlightProducer {
+  var _portNumber: AtomicInteger = new AtomicInteger(0)
 
+  def portNumber: AtomicInteger = _portNumber
   override def doAction(
       context: FlightProducer.CallContext,
       action: Action,
@@ -42,9 +51,14 @@ private class AmberProducer(
             )
           case payload =>
             throw new RuntimeException(s"not supported payload $payload")
-
         }
         listener.onNext(new Result("ack".getBytes))
+        listener.onCompleted()
+      case "handshake" =>
+        val strPortNumber: String = new String(action.getBody, Charset.forName("UTF-8"))
+        // Receive the port number from Python and put it into promise
+        promise.setValue(strPortNumber.toInt)
+        listener.onNext(new Result("ok".getBytes))
         listener.onCompleted()
       case _ => throw new NotImplementedError()
     }
@@ -88,18 +102,25 @@ private class AmberProducer(
 }
 
 class PythonProxyServer(
-    portNumber: Int,
     controlOutputPort: NetworkOutputPort[ControlPayload],
     dataOutputPort: NetworkOutputPort[DataPayload],
-    val actorId: ActorVirtualIdentity
+    val actorId: ActorVirtualIdentity,
+    promise: Promise[Int]
 ) extends Runnable
     with AutoCloseable
     with AmberLogging {
+  private lazy val portNumber: AtomicInteger = new AtomicInteger(getFreeLocalPort)
+  def getPortNumber: AtomicInteger = portNumber
 
   val allocator: BufferAllocator =
     new RootAllocator().newChildAllocator("flight-server", 0, Long.MaxValue);
-  val location: Location = Location.forGrpcInsecure("localhost", portNumber)
-  val producer: FlightProducer = new AmberProducer(controlOutputPort, dataOutputPort)
+
+  val producer: FlightProducer = new AmberProducer(controlOutputPort, dataOutputPort, promise)
+
+  val location: Location = (() => {
+    Location.forGrpcInsecure("localhost", portNumber.intValue())
+  })()
+
   val server: FlightServer = FlightServer.builder(allocator, location, producer).build()
 
   override def run(): Unit = {
@@ -111,4 +132,25 @@ class PythonProxyServer(
     AutoCloseables.close(server, allocator)
   }
 
+  /**
+    * Get a random free port.
+    *
+    * @return The port number.
+    * @throws IOException  , might happen when getting a free port.
+    */
+  @throws[IOException]
+  private def getFreeLocalPort: Int = {
+    var s: ServerSocket = null
+    try {
+      // ServerSocket(0) results in availability of a free random port
+      s = new ServerSocket(0)
+      s.getLocalPort
+    } catch {
+      case e: Exception =>
+        throw new RuntimeException(e)
+    } finally {
+      assert(s != null)
+      s.close()
+    }
+  }
 }
