@@ -1,13 +1,19 @@
 package edu.uci.ics.amber.engine.architecture.worker
 
 import edu.uci.ics.amber.engine.architecture.deploysemantics.layer.OpExecConfig
+import edu.uci.ics.amber.engine.common.virtualidentity.util.SOURCE_STARTER_OP
 import edu.uci.ics.amber.engine.common.virtualidentity.{ActorVirtualIdentity, LinkIdentity}
 
 import scala.collection.mutable
 
 class UpstreamLinkStatus(opExecConfig: OpExecConfig) {
 
-  val allUpstreamLinkIds: Set[LinkIdentity] = opExecConfig.inputToOrdinalMapping.keySet
+  private val allUpstreamLinkIds: Set[LinkIdentity] = {
+    if (opExecConfig.isSourceOperator)
+      Set(LinkIdentity(SOURCE_STARTER_OP, opExecConfig.id)) // special case for source operator
+    else
+      opExecConfig.inputToOrdinalMapping.keySet
+  }
 
   /**
     * The scheduler may not schedule the entire workflow at once. Consider a 2-phase hash join where the first
@@ -17,14 +23,18 @@ class UpstreamLinkStatus(opExecConfig: OpExecConfig) {
     * the build part completes. Therefore, we have a `allUpstreamLinkIds` to track the number of actual upstream
     * links that a worker receives data from.
     */
-  val upstreamMap =
+  private val upstreamMap =
     new mutable.HashMap[LinkIdentity, Set[ActorVirtualIdentity]].withDefaultValue(Set())
-  val upstreamMapReverse =
+  private val upstreamMapReverse =
     new mutable.HashMap[ActorVirtualIdentity, LinkIdentity]
   private val endReceivedFromWorkers = new mutable.HashSet[ActorVirtualIdentity]
   private val completedLinkIds = new mutable.HashSet[LinkIdentity]()
 
   def registerInput(identifier: ActorVirtualIdentity, input: LinkIdentity): Unit = {
+    assert(
+      allUpstreamLinkIds.contains(input),
+      "unexpected input link " + input + " for operator " + opExecConfig.id
+    )
     upstreamMap.update(input, upstreamMap(input) + identifier)
     upstreamMapReverse.update(identifier, input)
   }
