@@ -9,6 +9,7 @@ import edu.uci.ics.texera.web.model.jooq.generated.tables.FileOfProject;
 import edu.uci.ics.texera.web.model.jooq.generated.tables.FileOfWorkflow;
 import edu.uci.ics.texera.web.model.jooq.generated.tables.Project;
 import edu.uci.ics.texera.web.model.jooq.generated.tables.ProjectUserAccess;
+import edu.uci.ics.texera.web.model.jooq.generated.tables.PublicProject;
 import edu.uci.ics.texera.web.model.jooq.generated.tables.User;
 import edu.uci.ics.texera.web.model.jooq.generated.tables.UserConfig;
 import edu.uci.ics.texera.web.model.jooq.generated.tables.UserFileAccess;
@@ -34,7 +35,7 @@ import org.jooq.impl.SchemaImpl;
 @SuppressWarnings({ "all", "unchecked", "rawtypes" })
 public class TexeraDb extends SchemaImpl {
 
-    private static final long serialVersionUID = -1778414185;
+    private static final long serialVersionUID = -804967438;
 
     /**
      * The reference instance of <code>texera_db</code>
@@ -65,6 +66,11 @@ public class TexeraDb extends SchemaImpl {
      * The table <code>texera_db.project_user_access</code>.
      */
     public final ProjectUserAccess PROJECT_USER_ACCESS = edu.uci.ics.texera.web.model.jooq.generated.tables.ProjectUserAccess.PROJECT_USER_ACCESS;
+
+    /**
+     * The table <code>texera_db.public_project</code>.
+     */
+    public final PublicProject PUBLIC_PROJECT = edu.uci.ics.texera.web.model.jooq.generated.tables.PublicProject.PUBLIC_PROJECT;
 
     /**
      * The table <code>texera_db.user</code>.
@@ -138,6 +144,7 @@ public class TexeraDb extends SchemaImpl {
             FileOfWorkflow.FILE_OF_WORKFLOW,
             Project.PROJECT,
             ProjectUserAccess.PROJECT_USER_ACCESS,
+            PublicProject.PUBLIC_PROJECT,
             User.USER,
             UserConfig.USER_CONFIG,
             UserFileAccess.USER_FILE_ACCESS,
