@@ -208,7 +208,7 @@ class DashboardResource {
           WORKFLOW.LAST_MODIFIED_TIME,
           WORKFLOW_USER_ACCESS.PRIVILEGE,
           WORKFLOW_OF_USER.UID,
-          USER.NAME.as("userName"),
+          USER.NAME,
           // project attributes: 3 columns
           DSL.inline(null, classOf[UInteger]).as("pid"),
           DSL.inline(null, classOf[UInteger]).as("owner_id"),
@@ -268,9 +268,9 @@ class DashboardResource {
         DSL.inline(null, classOf[UserFileAccessPrivilege]).as("user_file_access")
       )
       .from(PROJECT)
-      .join(USER)
-      .on(PROJECT.OWNER_ID.eq(USER.UID))
-      .where(PROJECT.OWNER_ID.eq(user.getUid))
+      .leftJoin(PROJECT_USER_ACCESS)
+      .on(PROJECT_USER_ACCESS.PID.eq(PROJECT.PID))
+      .where(PROJECT_USER_ACCESS.UID.eq(user.getUid))
       .and(
         projectMatchQuery
       )
@@ -620,10 +620,12 @@ class DashboardResource {
             if (resourceType == "file") {
               DashboardFile(
                 record.into(USER).getEmail,
-                record.get(
-                  "user_file_access",
-                  classOf[UserFileAccessPrivilege]
-                ) == UserFileAccessPrivilege.WRITE,
+                record
+                  .get(
+                    "user_file_access",
+                    classOf[UserFileAccessPrivilege]
+                  )
+                  .toString,
                 record.into(FILE).into(classOf[File])
               )
             } else {
