@@ -2,7 +2,7 @@ package edu.uci.ics.texera.workflow.operators.sortPartitions
 
 import com.fasterxml.jackson.annotation.{JsonProperty, JsonPropertyDescription}
 import com.google.common.base.Preconditions
-import com.kjetland.jackson.jsonSchema.annotations.JsonSchemaTitle
+import com.kjetland.jackson.jsonSchema.annotations.{JsonSchemaInject, JsonSchemaTitle}
 import edu.uci.ics.amber.engine.architecture.deploysemantics.layer.OpExecConfig
 import edu.uci.ics.texera.workflow.common.metadata.annotations.AutofillAttributeName
 import edu.uci.ics.texera.workflow.common.metadata.{
@@ -15,6 +15,15 @@ import edu.uci.ics.texera.workflow.common.operators.OperatorDescriptor
 import edu.uci.ics.texera.workflow.common.tuple.schema.{OperatorSchemaInfo, Schema}
 import edu.uci.ics.texera.workflow.common.workflow.RangePartition
 
+@JsonSchemaInject(json = """
+{
+  "attributeTypeRules": {
+    "sortAttributeName":{
+      "enum": ["integer", "long", "double"]
+    }
+  }
+}
+""")
 class SortPartitionsOpDesc extends OperatorDescriptor {
 
   @JsonProperty(required = true)
