@@ -30,8 +30,8 @@ class ScatterplotVizOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
     .add(new Attribute("field2", AttributeType.DOUBLE), 32.33)
     .build()
 
-  //another tuple with slightly different latitude value to test pixel based visualization
-  val extratuple: Tuple = Tuple
+  // another tuple with slightly different latitude value to test pixel based visualization
+  val extraTuple: Tuple = Tuple
     .newBuilder(tupleSchema)
     .add(new Attribute("field1", AttributeType.DOUBLE), 73.142)
     .add(new Attribute("field2", AttributeType.DOUBLE), 32.22)
@@ -49,14 +49,15 @@ class ScatterplotVizOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
     scatterplotOpExec.open()
   }
 
-  it should "output the correct schema for geometric for the frontend to render it and process only one tuple because the other falls on the same pixel" in {
+  it should "output the correct schema for geometric for the frontend to render it and process only one tuple " +
+    "because the other falls on the same pixel" in {
     val processedTuple: Tuple =
       scatterplotOpExec.processTexeraTuple(Left(tuple), 0, null, null).next()
     assert(processedTuple.getField("xColumn").asInstanceOf[Double] == 73.142)
     assert(processedTuple.getField("yColumn").asInstanceOf[Double] == 32.33)
     val outputTuples: List[Tuple] =
-      scatterplotOpExec.processTexeraTuple(Left(extratuple), 0, null, null).toList
-    assert(outputTuples.size == 0)
+      scatterplotOpExec.processTexeraTuple(Left(extraTuple), 0, null, null).toList
+    assert(outputTuples.isEmpty)
   }
 
   it should "process the both points with the correct schema when the type is not geometric" in {
@@ -71,7 +72,7 @@ class ScatterplotVizOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
     assert(processedTuple.getField("field1").asInstanceOf[Double] == 73.142)
     assert(processedTuple.getField("field2").asInstanceOf[Double] == 32.33)
     val processedAnotherTuple: Tuple =
-      scatterplotOpExec.processTexeraTuple(Left(extratuple), 0, null, null).next()
+      scatterplotOpExec.processTexeraTuple(Left(extraTuple), 0, null, null).next()
     assert(processedAnotherTuple.getField("field1").asInstanceOf[Double] == 73.142)
     assert(processedAnotherTuple.getField("field2").asInstanceOf[Double] == 32.22)
   }
@@ -79,7 +80,7 @@ class ScatterplotVizOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
   it should "handle last tuple correctly" in {
     val outputTuples: List[Tuple] =
       scatterplotOpExec.processTexeraTuple(Right(InputExhausted()), 0, null, null).toList
-    assert(outputTuples.size == 0)
+    assert(outputTuples.isEmpty)
   }
 
   after {
