@@ -15,7 +15,7 @@ import { ShareAccessComponent } from "../../share-access/share-access.component"
 export class UserFileListItemComponent {
   private _entry?: DashboardFile = {
     ownerEmail: "jingchf@uci.edu",
-    writeAccess: true,
+    accessLevel: "WRITE",
     file: {
       ownerUid: 1,
       fid: 2,
@@ -41,6 +41,7 @@ export class UserFileListItemComponent {
   editingName = false;
   editingDescription = false;
   @Output() deleted = new EventEmitter<void>();
+  @Output() refresh = new EventEmitter<void>();
 
   constructor(
     private modalService: NgbModal,
@@ -87,8 +88,12 @@ export class UserFileListItemComponent {
 
   public onClickOpenShareAccess(): void {
     const modalRef = this.modalService.open(ShareAccessComponent);
+    modalRef.componentInstance.writeAccess = this.entry.accessLevel === "WRITE";
     modalRef.componentInstance.type = "file";
     modalRef.componentInstance.id = this.entry.file.fid;
+    modalRef.closed.pipe(untilDestroyed(this)).subscribe(_ => {
+      this.refresh.emit();
+    });
   }
 
   public downloadFile(): void {
