@@ -5,6 +5,8 @@ import { NotificationService } from "src/app/common/service/notification/notific
 import { UntilDestroy, untilDestroyed } from "@ngneat/until-destroy";
 import { ShareAccessComponent } from "../../share-access/share-access.component";
 import { NgbModal } from "@ng-bootstrap/ng-bootstrap";
+import { PublicProjectService } from "../../../service/public-project/public-project.service";
+import { UserService } from "../../../../../common/service/user/user.service";
 
 @UntilDestroy()
 @Component({
@@ -27,13 +29,16 @@ export class UserProjectListItemComponent implements OnInit {
     this._entry = value;
   }
   @Output() deleted = new EventEmitter<void>();
+  @Output() refresh = new EventEmitter<void>();
   @Input() editable = false;
   @Input() uid: number | undefined;
   editingColor = false;
   editingName = false;
   editingDescription = false;
-  descriptionCollapsed = false;
+  descriptionCollapsed = true;
   color = "#ffffff";
+  isAdmin: boolean = false;
+  isPublic: boolean = false;
   /** To make sure info remains visible against white background */
   get lightColor() {
     return UserProjectService.isLightColor(this.color);
@@ -42,13 +47,23 @@ export class UserProjectListItemComponent implements OnInit {
   constructor(
     private userProjectService: UserProjectService,
     private notificationService: NotificationService,
-    private modalService: NgbModal
-  ) {}
+    private modalService: NgbModal,
+    private userService: UserService,
+    private publicProjectService: PublicProjectService
+  ) {
+    this.isAdmin = this.userService.isAdmin();
+  }
 
   ngOnInit(): void {
     if (this.entry.color) {
       this.color = this.entry.color;
     }
+    this.publicProjectService
+      .getType(this.entry.pid)
+      .pipe(untilDestroyed(this))
+      .subscribe(type => {
+        this.isPublic = type === "Public";
+      });
   }
 
   updateProjectColor(): void {
@@ -72,15 +87,9 @@ export class UserProjectListItemComponent implements OnInit {
     this.userProjectService
       .updateProjectColor(this.entry.pid, color)
       .pipe(untilDestroyed(this))
-      .subscribe({
-        next: () => {
-          this.color = color;
-          this.entry = { ...this.entry, color: color };
-        },
-        error: (err: unknown) => {
-          // @ts-ignore
-          this.notificationService.error(err.error.message);
-        },
+      .subscribe(() => {
+        this.color = color;
+        this.entry = { ...this.entry, color: color };
       });
   }
 
@@ -90,15 +99,9 @@ export class UserProjectListItemComponent implements OnInit {
     this.userProjectService
       .deleteProjectColor(this.entry.pid)
       .pipe(untilDestroyed(this))
-      .subscribe({
-        next: _ => {
-          this.color = "#ffffff"; // reset color wheel
-          this.entry = { ...this.entry, color: null };
-        },
-        error: (err: unknown) => {
-          // @ts-ignore
-          this.notificationService.error(err.error.message);
-        },
+      .subscribe(() => {
+        this.color = "#ffffff"; // reset color wheel
+        this.entry = { ...this.entry, color: null };
       });
   }
 
@@ -110,18 +113,12 @@ export class UserProjectListItemComponent implements OnInit {
       this.userProjectService
         .updateProjectName(this.entry.pid, name)
         .pipe(untilDestroyed(this))
-        .subscribe({
-          next: () => {
-            if (!this.entry) {
-              throw new Error("entry property must be provided to UserProjectListItemComponent.");
-            }
-            this.editingName = false;
-            this.entry.name = name;
-          },
-          error: (err: unknown) => {
-            // @ts-ignore
-            this.notificationService.error(err.error.message);
-          },
+        .subscribe(() => {
+          if (!this.entry) {
+            throw new Error("entry property must be provided to UserProjectListItemComponent.");
+          }
+          this.editingName = false;
+          this.entry.name = name;
         });
     }
   }
@@ -137,27 +134,33 @@ export class UserProjectListItemComponent implements OnInit {
     this.userProjectService
       .updateProjectDescription(this.entry.pid, description)
       .pipe(untilDestroyed(this))
-      .subscribe({
-        next: () => {
-          this.entry.description = description;
-          this.notificationService.success(`Saved description for project: "${this.entry.name}".`);
-        },
-        error: (err: unknown) => {
-          // @ts-ignore
-          this.notificationService.error(err.error.message);
-        },
-      })
-      .add(() => {
+      .subscribe(() => {
+        this.entry.description = description;
+        this.notificationService.success(`Saved description for project: "${this.entry.name}".`);
         this.editingDescription = false;
       });
   }
 
-  /**
-   * open the Modal based on the workflow clicked on
-   */
-  public onClickOpenShareAccess(id: number): void {
+  public onClickOpenShareAccess(): void {
     const modalRef = this.modalService.open(ShareAccessComponent);
+    modalRef.componentInstance.writeAccess = this.entry.accessLevel === "WRITE";
     modalRef.componentInstance.type = "project";
-    modalRef.componentInstance.id = id;
+    modalRef.componentInstance.id = this.entry.pid;
+    modalRef.closed.pipe(untilDestroyed(this)).subscribe(_ => {
+      this.refresh.emit();
+    });
+  }
+  public visibilityChange(): void {
+    if (this.isPublic) {
+      this.publicProjectService
+        .makePrivate(this.entry.pid)
+        .pipe(untilDestroyed(this))
+        .subscribe(() => this.ngOnInit());
+    } else {
+      this.publicProjectService
+        .makePublic(this.entry.pid)
+        .pipe(untilDestroyed(this))
+        .subscribe(() => this.ngOnInit());
+    }
   }
 }
