@@ -1,7 +1,7 @@
 import { JSONSchema7, JSONSchema7Definition } from "json-schema";
 
 export const hideTypes = ["regex", "equals"] as const;
-export type HideType = typeof hideTypes[number];
+export type HideType = (typeof hideTypes)[number];
 
 export type AttributeTypeEnumRule = ReadonlyArray<string>;
 export type AttributeTypeConstRule = Readonly<{
